udmabuf: fix memfd_fcntl signature to satisfy kernel CFI
On sm8650 (OPPO 6.1.118, CONFIG_CFI_CLANG=y, CFI_PERMISSIVE unset) the
native-mode CREATE path hard-reset the entire SoC with no clean bark.
Bisected (via a temporary diag_stop module param, non-persistent) to the
F_GET_SEALS step in ukv_seals(): the indirect call through the
kallsyms-resolved memfd_fcntl pointer.

Root cause: the pointer was declared

  long (*)(struct file *, unsigned int, unsigned int)

but mm/memfd.c defines

  long memfd_fcntl(struct file *, unsigned int, unsigned long)

Under non-permissive kernel CFI, an indirect call whose type-id does not
match the target is a fatal violation, which on this SoC manifests as an
immediate hardware reset. The other gunyah modules' kallsyms-resolved
calls work precisely because their signatures match exactly.

Fix: declare the third argument as unsigned long. udmabuf_test at
64/256/512 MiB plus CREATE_LIST now all PASS with the device stable.

+++ udmabuf/udmabuf.c
@@ -166,7 +166,9 @@ MODULE_PARM_DESC(stat_failed, "create attempts this module rejected (read-only).
 /* ------------------------------------------------------------------------- */
 
 static unsigned long (*ukv_lookup_name)(const char *name);
-static long (*ukv_memfd_fcntl)(struct file *file, unsigned int cmd, unsigned int arg);
+/* Signature MUST match mm/memfd.c exactly: under CONFIG_CFI_CLANG a mismatched
+ * indirect-call type is a fatal CFI violation (arg is unsigned long, not int). */
+static long (*ukv_memfd_fcntl)(struct file *file, unsigned int cmd, unsigned long arg);
 static int (*ukv_lookup_size_offset)(unsigned long addr, unsigned long *symbolsize,
 				     unsigned long *offset);
